move proxy fixture to conftest
so it can be re-used
