+++ tests/test_proxy.py
@@ -138,36 +138,6 @@ def launch_backends():
         proc.wait()
 
 
-@pytest.fixture(
-    params=[
-        "no_auth_consul_proxy",
-        "auth_consul_proxy",
-        "no_auth_etcd_proxy",
-        "auth_etcd_proxy",
-        "file_proxy_toml",
-        "file_proxy_yaml",
-        "external_consul_proxy",
-        "auth_external_consul_proxy",
-        "external_etcd_proxy",
-        "auth_external_etcd_proxy",
-        "external_file_proxy_toml",
-        "external_file_proxy_yaml",
-    ]
-)
-def proxy(request):
-    """Parameterized fixture to run all the tests with every proxy implementation"""
-    proxy = request.getfixturevalue(request.param)
-    # wait for public endpoint to be reachable
-    asyncio.run(
-        exponential_backoff(
-            utils.check_host_up_http,
-            f"Proxy public url {proxy.public_url} cannot be reached",
-            url=proxy.public_url,
-        )
-    )
-    return proxy
-
-
 async def wait_for_services(urls):
     # Wait until traefik and the backend are ready
     await exponential_backoff(
